fixed terrain

+++ src/resources/primitives/meshes/GridMesh.hpp
@@ -26,12 +26,12 @@ namespace KCore {
         void createGeneralSurface(float width, float length, int segmentsX, int segmentsY);
 
         void createGeneralSurface(float width, float length, int segmentsX, int segmentsY,
-                                  const std::vector<uint16_t> &heights);
+                                  const std::vector<float> &heights);
 
         void makeHorizontalBorder(float width, int segments, float constraint,
-                                  const std::vector<uint16_t> &heights, bool downBorder);
+                                  const std::vector<float> &heights, bool downBorder);
 
         void makeVerticalBorder(float height, int segments, float constraint,
-                                const std::vector<uint16_t> &heights, bool rightBorder);
+                                const std::vector<float> &heights, bool rightBorder);
     };
 }
